requirements.txt
click
requests
simplejson
PyYAML